[ Refactor ] refactored CardInformationValidator class with the extension methods on CardInfromation directly, to mimimize code implementation, with remaining the main isValid class
